refactor: moving tests related to Authentication logic to Tests/Feature/Auth directory
